+++ lambdas/data-migration1/tests/test-collections.js
@@ -43,6 +43,7 @@ const generateFakeCollection = (params) => ({
 
 let collectionsModel;
 let rulesModel;
+const collectionOmitList = ['granuleId', 'granuleIdExtraction', 'sampleFileName', 'granuleIdValidationRegex', 'granuleIdExtractionRegex', 'ignoreFilesConfigForDiscovery', 'duplicateHandling', 'reportToEms', 'createdAt', 'updatedAt'];
 
 test.before(async (t) => {
   process.env.stackName = cryptoRandomString({ length: 10 });
@@ -101,16 +102,20 @@ test.serial('migrateCollectionRecord correctly migrates collection record', asyn
     .first();
 
   t.deepEqual(
-    omit(createdRecord, ['cumulusId']),
+    omit(createdRecord, ['cumulus_id']),
     omit(
       {
         ...fakeCollection,
-        granuleIdValidationRegex: fakeCollection.granuleId,
-        granuleIdExtractionRegex: fakeCollection.granuleIdExtraction,
+        sample_file_name: fakeCollection.sampleFileName,
+        granule_id_validation_regex: fakeCollection.granuleId,
+        granule_id_extraction_regex: fakeCollection.granuleIdExtraction,
+        duplicate_handling: fakeCollection.duplicateHandling,
+        report_to_ems: fakeCollection.reportToEms,
+        ignore_files_config_for_discovery: fakeCollection.ignoreFilesConfigForDiscovery,
         created_at: new Date(fakeCollection.createdAt),
         updated_at: new Date(fakeCollection.updatedAt),
       },
-      ['granuleId', 'granuleIdExtraction', 'createdAt', 'updatedAt']
+      collectionOmitList
     )
   );
 });
@@ -145,24 +150,25 @@ test.serial('migrateCollectionRecord handles nullable fields on source collectio
     .first();
 
   t.deepEqual(
-    omit(createdRecord, ['cumulusId']),
+    omit(createdRecord, ['cumulus_id']),
     omit(
       {
         ...fakeCollection,
-        granuleIdValidationRegex: fakeCollection.granuleId,
-        granuleIdExtractionRegex: fakeCollection.granuleIdExtraction,
+        sample_file_name: fakeCollection.sampleFileName,
+        granule_id_validation_regex: fakeCollection.granuleId,
+        granule_id_extraction_regex: fakeCollection.granuleIdExtraction,
         url_path: null,
         process: null,
-        ignoreFilesConfigForDiscovery: null,
+        ignore_files_config_for_discovery: null,
         meta: null,
         tags: null,
         created_at: new Date(fakeCollection.createdAt),
         updated_at: new Date(fakeCollection.updatedAt),
         // schema validation will add default values
-        duplicateHandling: 'error',
-        reportToEms: true,
+        duplicate_handling: 'error',
+        report_to_ems: true,
       },
-      ['granuleId', 'granuleIdExtraction', 'createdAt', 'updatedAt']
+      collectionOmitList
     )
   );
 });