+++ packages/db/src/index.ts
@@ -1,10 +1,17 @@
+export {
+  createTestDatabase, deleteTestDatabase,
+} from './database';
 export { getKnexClient } from './connection';
 export { getKnexConfig, localStackConnectionEnv } from './config';
 export { doesRecordExist, isRecordDefined } from './database';
 export { tableNames } from './tables';
 export {
-  AsyncOperationRecord,
-  CollectionRecord,
+  PostgresAsyncOperation,
+  PostgresAsyncOperationRecord,
+  PostgresCollection,
+  PostgresCollectionRecord,
   ExecutionRecord,
   ProviderRecord,
 } from './types';
+export { translateApiAsyncOperationToPostgresAsyncOperation } from './async_operations';
+export { translateApiCollectionToPostgresCollection } from './collections';