+++ packages/integration-tests/index.js
@@ -60,8 +60,6 @@ const lambdaStep = new LambdaStep();
  *   number of retries has been exceeded.
  *
  * @param {Object} params - params
- * @param {string} params.TableName - the name of the AsyncOperations DynamoDB
- *   table
  * @param {string} params.id - the id of the AsyncOperation
  * @param {string} params.status - the status to wait for
  * @param {number} params.retries - the number of times to retry Default: 10